gnp.py: Fix dir copy bug in backup op (#11)
Don't need to copy .webgpu_typescript from source code, which is
automatically created when running webgpu cts. But it fails to create
due to webgpu cts source files are not copies to correct dir.

Dir paths may or may not have separators, but their dirname got
by os.path.dirname are different, for example
os.path.dirname('/path/to/src/') is '/path/to/src', but
os.path.dirname('/path/to/src') is 'path/to'.