gputest.py: Update trace test expectations (#28)
VideoPathTraceTest and OverlayModeTraceTest in trace_test are expected
failures on Intel devices. We need to update the trace test expectations
as we did in WebGPU CTS.

Not need to check Intel GPU name when updating the expectations, we only
run gputest.py on Intel devices, and the update won't break the tests on
non-Intel devices.

Also updated the helper function name in chrome drop since it's renamed.